+++ app/(default)/tools/participants/[studyKey]/participants/exporter/_components/ParticipantDownloader.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import LoadingButton from '@/components/LoadingButton';
+import LoadingButton from '@/components/loading-button';
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import { Input } from '@/components/ui/input';
