+++ app/(default)/tools/participants/[studyKey]/responses/_components/DeleteResponsesDialog.tsx
@@ -1,5 +1,5 @@
 'use client'
-import LoadingButton from '@/components/LoadingButton';
+import LoadingButton from '@/components/loading-button';
 import { Button } from '@/components/ui/button';
 import { Dialog, DialogClose, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/dialog';
 import { AlertTriangle } from 'lucide-react';
