+++ app/(default)/tools/participants/[studyKey]/responses/exporter/_components/SurveyInfoDownloader.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import LoadingButton from "@/components/LoadingButton";
+import LoadingButton from "@/components/loading-button";
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
 import { AlertTriangle, CheckCircle, Download } from "lucide-react";
