+++ app/(default)/tools/study-configurator/[studyKey]/settings/code-lists/_components/add-code-list-entries.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import LoadingButton from '@/components/LoadingButton';
+import LoadingButton from '@/components/loading-button';
 import { Button } from '@/components/ui/button';
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
 import { PlusIcon } from 'lucide-react';
