+++ app/(default)/tools/study-configurator/new/CreateStudyForm.tsx
@@ -10,7 +10,7 @@ import { useForm } from "react-hook-form"
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Switch } from '@/components/ui/switch';
-import LoadingButton from '@/components/LoadingButton';
+import LoadingButton from '@/components/loading-button';
 import { toast } from 'sonner';
 import { createStudy } from '@/actions/study/create';
 import { useRouter } from 'next/navigation';
