+++ app/(default)/tools/user-management/management-users/[userID]/_components/AddPermissionDialog.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useTransition } from 'react';
-import LoadingButton from '@/components/LoadingButton';
+import LoadingButton from '@/components/loading-button';
 import { Button } from '@/components/ui/button';
 import { Dialog, DialogClose, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
 import { Plus } from 'lucide-react';
