+++ app/(default)/tools/user-management/management-users/[userID]/_components/UpdatePermissionLimiterDialog.tsx
@@ -9,7 +9,7 @@ import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, For
 import { Textarea } from '@/components/ui/textarea';
 import { Dialog, DialogClose, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
-import LoadingButton from '@/components/LoadingButton';
+import LoadingButton from '@/components/loading-button';
 import { Badge } from '@/components/ui/badge';
 import { ManagementUserPermission } from '@/lib/data/userManagementAPI';
