+++ app/(default)/tools/user-management/service-accounts/[serviceAccountID]/_components/service-account-infos.tsx
@@ -1,5 +1,5 @@
 'use client';
-import LoadingButton from '@/components/LoadingButton';
+import LoadingButton from '@/components/loading-button';
 import { Card, CardContent, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
