+++ components/survey-renderer/SurveySingleItemView/ResponseComponent/InputTypes/code-validator.tsx
@@ -10,7 +10,7 @@ import { Button } from "@/components/ui/button";
 import { AlertTriangleIcon, CheckCircle2Icon, Pen } from "lucide-react";
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
-import LoadingButton from "@/components/LoadingButton";
+import LoadingButton from "@/components/loading-button";
 
 
 type CodeValidatorProps = CommonResponseComponentProps
